refactor(upgrade): split hosted transaction durability
